Nowcast quarter changed to be the first incomplete/unbalanced quarter.
Output changed such that all analysis should employed using the SSM and
not on the draws of the data from the posterior data distribution. This
should keep all the analysis mutually consistent.

SSM_fixed object added that assumes you observe the quarterly data at a
monthly frequency at the median of the posterior data distribution. This
abstracts away from the uncertainty originating from having to estimate
the higher frequency data and leaves over the uncertainty related to the
nowcast/forecast of the variables of interest.

I intend to focus the analysis of the model results on the difference
between a forecast without any intra-quarter information with the
various vintages respectively. This will hopefully make the usefullness
of the "nowcast" more readily interpretable.